Move decorators to app/decorators
