--- app/controllers/spree/frontend/checkout_controller_decorator.rb
@@ -1,12 +1,0 @@
-Spree::CheckoutController.class_eval do
-  rescue_from SpreeTaxCloud::Error do |exception|
-    flash[:error] = exception.message
-    redirect_to checkout_state_path(:address)
-  end
-
-  rescue_from TaxCloud::Errors::ApiError do |exception|
-    exception_message = exception.problem
-    flash[:error] = I18n.t('spree.address_verification_failed') + (exception_message ? ": #{exception_message}" : '')
-    redirect_to checkout_state_path(:address)
-  end
-end
