+++ lib/spree_tax_cloud.rb
@@ -1,6 +1,7 @@
 require 'spree_api'
 require 'spree_backend'
 require 'spree_core'
+require 'solidus_support'
 require 'spree_tax_cloud/engine'
 require 'spree_tax_cloud/error'
 require 'tax_cloud'
